Extracting new test start time generation generation to a helper function
