+++ src/shared/helpers/__tests__/get-new-test-start-time.ts
@@ -1,0 +1,8 @@
+
+describe('getNewTestStartTime helper function', () => {
+
+  it('should return the correct date', () => {
+    expect(true).toBe(false);
+  });
+
+});
